# backend/docs/swagger.yaml
definitions:
...
  handler.APIResponse:
    properties:
      data: {}
      msg:
        example: ok
        type: string
    type: object
...
  handler.UserListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.UserResponse'
        type: array
      total:
        example: 100
        type: integer
    type: object
  handler.UserResponse:
    properties:
      age:
        example: 36
        type: integer
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      b_end_identities:
        example:
        - staff
        - station_manager
        items:
          type: string
        type: array
      birth_date:
        example: "1990-01-01"
        type: string
      c_end_identities:
        example:
        - elderly
        items:
          type: string
        type: array
      created_at:
        type: string
      email:
        example: zhangsan@example.com
        type: string
      gender:
        example: male
        type: string
      id:
        example: 1
        type: integer
      id_card_hash:
        example: 2f0ab5...
        type: string
      id_card_masked:
        example: 1101**********1234
        type: string
      id_card_token:
        example: eyJ1aWQiOjF9.xxx
        type: string
      name:
        example: 张三
        type: string
      phone:
        example: "13800000001"
        type: string
      primary_identity:
        example: staff
        type: string
      station_id:
        example: 1
        type: integer
      status:
        example: active
        type: string
      updated_at:
        type: string
    type: object
...
  handler.userCreateRequest:
    properties:
      email:
        type: string
      identity_type:
        description: 身份类型
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      station_id:
        type: integer
      status:
        type: string
    required:
    - identity_type
    - password
    - phone
    type: object
  handler.userUpdateRequest:
    properties:
      avatar:
        type: string
      birth_date:
        description: 格式：2006-01-02
        type: string
      email:
        type: string
      gender:
        type: string
      id_card:
        type: string
      id_card_hash:
        type: string
      id_card_token:
        type: string
      name:
        type: string
      password:
        type: string
      station_id:
        type: integer
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /b/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传文件到存储服务，B端和C端有不同的限制
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 模块名称
        in: query
        name: module
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功，返回url和key
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 文件缺失或不符合要求
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 上传失败
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 上传文件
      tags:
      - c_upload
      - b_upload
  /b/users:
    get:
      consumes:
      - application/json
      description: 分页获取B端用户列表，支持筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 身份筛选(elderly/family/staff/station_manager/admin)
        in: query
        name: role
        type: string
      - description: 状态筛选(active/inactive/suspended)
        in: query
        name: status
        type: string
      - description: 站点ID筛选
        in: query
        name: station_id
        type: integer
      - description: 关键词搜索(姓名/手机号)
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.UserListResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取B端用户列表
      tags:
      - b_user
    post:
      consumes:
      - application/json
      description: 创建新的B端工作人员账号
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.userCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.UserResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 创建B端用户
      tags:
      - b_user
  /b/users/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取B端用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.UserResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取B端用户详情
      tags:
      - b_user
    put:
      consumes:
      - application/json
      description: 更新指定B端用户的信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.userUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.UserResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 更新B端用户
      tags:
      - b_user
...
  /c/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传文件到存储服务，B端和C端有不同的限制
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 模块名称
        in: query
        name: module
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功，返回url和key
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 文件缺失或不符合要求
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 上传失败
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 上传文件
      tags:
      - c_upload
      - b_upload
swagger: "2.0"